Add static_cast

+++ Source/GMTK_Jam_2024/Actors/Scales.cpp
@@ -111,8 +111,8 @@ void AScales::CalculateBalance()
 		return;
 	}
 
-	const float RightBowlWeight = RightBowl->GetWeight();
-	const float LeftBowlWeight = LeftBowl->GetWeight();
+	const float RightBowlWeight = static_cast<float>(RightBowl->GetWeight());
+	const float LeftBowlWeight = static_cast<float>(LeftBowl->GetWeight());
 	WeightBalance = LeftBowlWeight <= 0.f ? 0.f : RightBowlWeight / LeftBowlWeight - 1.f;
 }
